features: create game

+++ src/features/quiz/application/admin-use-cases/delete-question.use-case.ts
@@ -1,8 +1,8 @@
 import { CommandHandler, ICommandHandler } from "@nestjs/cqrs";
 import { NotFoundException } from "@nestjs/common";
+import { Logger } from "@nestjs/common";
 
 import { QuizGameQuestion } from "../../domain/quiz-game-question.entity";
-
 import { SaQuizTypeormRepository } from "../../infra/sa-quiz-typeorm.repository";
 
 export class DeleteQuestionCommand {
@@ -13,6 +13,8 @@ export class DeleteQuestionCommand {
 
 @CommandHandler(DeleteQuestionCommand)
 export class DeleteQuestionUseCase implements ICommandHandler<DeleteQuestionCommand> {
+  private readonly logger = new Logger(DeleteQuestionUseCase.name);
+
   constructor(
     private readonly saQuizTypeormRepository: SaQuizTypeormRepository
   ) {}
@@ -20,10 +22,20 @@ export class DeleteQuestionUseCase implements ICommandHandler<DeleteQuestionComm
   async execute(command: DeleteQuestionCommand): Promise<void> {
     const { id } = command;
 
-    const res = await this.saQuizTypeormRepository.deleteQuestion(id)
+    const question = await this.saQuizTypeormRepository.questionIsExist(id);
 
-    if (!res) {
-      throw new NotFoundException(`Blog with id ${id} not found`);
+    if (!question) {
+      throw new NotFoundException(`Question with id ${id} not found`);
     }
+
+    this.logger.log(`Deleting question with id ${id}`);
+
+    const deleteResult = await this.saQuizTypeormRepository.deleteQuestion(id);
+
+    if (!deleteResult) {
+      throw new NotFoundException(`Question with id ${id} not deleted`);
+    }
+
+    this.logger.log(`Question with id ${id} successfully deleted`);
   }
 }
